Add vhost user library
This repo provides the APIs to create and destroy vhost user backend
device, provides the handlers to process vhost user messages, and
provides the APIs to pop and push the virtqueue requests.

Note that, VHOST_USER_F_PROTOCOL_FEATURES is not supported in this
version. Will add the support in later version if there is requirement
for it.

Change-Id: I34af56a11c3e8de979b68ea034879e9495841e80
Signed-off-by: Conghui Chen <quic_conghuic@quicinc.com>