clear the nr_vring during device deinit
The second time GVM starts, the nr_vring for virtio device still keep the
value set by the first launch. When the virtio device is not fully initiated,
and the deinit is triggered, it will try to free the virtqueue which is not
created.

so, clear the nr_vring during deinit to avoid such error.

Change-Id: Ia76c3937ce01f0f3720960dffac76c8064efe995
Signed-off-by: Conghui Chen <quic_conghuic@quicinc.com>